model-server/requirements.txt
compressai==1.2.1
grpcio==1.51.1
Pillow==9.4.0
protobuf==4.21.12
pytorch_msssim==0.2.1
torch==1.12.1
torchvision==0.13.1
waitress==2.1.2
openpyxl~=3.1.0
pandas~=1.5.2